requirements.txt
flask==2.3.3
flask-socketio==5.3.6
requests==2.31.0
pandas==2.2.2
scikit-learn==1.5.1
python-dotenv==1.0.1